pyproject.toml
[tool.poetry]
name = "fusm_data"
version = "0.1.0"
description = "Capa de datos para los sistemas del FabLab UTFSM"
authors = ["Pablo Albornoz"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3"
djangorestframework = "^3.11.0"
psycopg2-binary = "^2.8.4"
Pillow = "^7.0.0"
pyyaml = "^5.3"
drf-yasg = "^1.17.0"
drf-extensions = "^0.6.0"

[tool.poetry.dev-dependencies]
pylint = "pylint-django"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
